+++ ucoin/src/ln/ln_node.c
@@ -63,6 +63,10 @@ bool ln_node_init(ln_node_t *node, const char *pWif, const char *pNodeName, uint
         anno.timestamp = (uint32_t)time(NULL);
         anno.p_my_node = &node->keys;
         anno.p_alias = node->alias;
+        anno.rgbcolor[0] = 0;
+        anno.rgbcolor[1] = 0;
+        anno.rgbcolor[2] = 0;
+        memcpy(&anno.addr, &node->addr, sizeof(ln_nodeaddr_t));
         ret = ln_msg_node_announce_create(&buf_node, &anno);
         if (!ret) {
             goto LABEL_EXIT;
@@ -172,6 +176,10 @@ bool HIDDEN ln_node_recv_node_announcement(ln_self_t *self, const uint8_t *pData
         buf_ann.buf = (CONST_CAST uint8_t *)pData;
         buf_ann.len = Len;
         ret = ln_db_save_anno_node(&buf_ann, ln_their_node_id(self), node_pub);
+
+        if (ret) {
+            (*self->p_callback)(self, LN_CB_NODE_ANNO_RECV, &ann);
+        }
     }
     ucoin_buf_free(&buf_old);